pjt_backend/requirements.txt
asgiref==3.8.1
attrs==24.2.0
backports.tarfile==1.2.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
contourpy==1.3.0
cryptography==43.0.3
cycler==0.12.1
defusedxml==0.7.1
dj-rest-auth==7.0.0
Django==4.2.16
django-allauth==0.61.1
django-cors-headers==4.6.0
django-environ==0.11.2
django-extensions==3.2.3
django-rest-framework==0.1.0
djangorestframework==3.15.2
docutils==0.21.2
drf-spectacular==0.27.2
fonttools==4.55.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
inflection==0.5.1
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keyring==25.5.0
kiwisolver==1.4.7
markdown-it-py==3.0.0
matplot==0.1.9
matplotlib==3.9.2
mdurl==0.1.2
more-itertools==10.5.0
nh3==0.2.18
numpy==2.0.2
oauthlib==3.2.2
packaging==24.2
pandas==2.2.3
pillow==11.0.0
pkginfo==1.10.0
pycparser==2.22
Pygments==2.18.0
PyJWT==2.10.0
pyloco==0.0.139
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytz==2024.2
pywin32-ctypes==0.2.3
PyYAML==6.0.2
readme_renderer==44.0
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.9.4
rpds-py==0.21.0
SimpleWebSocketServer==0.1.2
six==1.16.0
sqlparse==0.5.2
twine==5.1.1
typing==3.7.4.3
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
ushlex==0.99.1
websocket-client==1.8.0
zipp==3.21.0
